# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Azubi Commerce API",
        "description": "Azubi Commerce API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "local server"
        },
        {
            "url": "https://example.com",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Returns all products",
                "description": "Returns a all products",
                "operationId": "getInventory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Save product information",
                "description": "Saves products to the database",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Product name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Product Name"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Product price",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 29.99
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Product category",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Category Name"
                        }
                    },
                    {
                        "name": "thumbnail",
                        "in": "query",
                        "description": "Thumbnail image URL",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "image_thumbnail.jpg"
                        }
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Mobile image URL",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "image_mobile.jpg"
                        }
                    },
                    {
                        "name": "desktop",
                        "in": "query",
                        "description": "Desktop image URL",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "image_desktop.jpg"
                        }
                    },
                    {
                        "name": "tablet",
                        "in": "query",
                        "description": "Tablet image URL",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "image_tablet.jpg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "results": {
                                            "type": "string",
                                            "example": "Product created successfully with respective images"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Returns searched product",
                "description": "Returns a product by name",
                "operationId": "3594217dfd087ad7888b1099abc150ed",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "name of product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get a product by its ID",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "results": {
                                            "type": "object",
                                            "example": "successfully fetched"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "Update an existing product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Product Name"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 39.99
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Updated Category Name"
                                    },
                                    "thumbnail": {
                                        "type": "string",
                                        "example": "updated_thumbnail.jpg"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "updated_mobile.jpg"
                                    },
                                    "desktop": {
                                        "type": "string",
                                        "example": "updated_desktop.jpg"
                                    },
                                    "tablet": {
                                        "type": "string",
                                        "example": "updated_tablet.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "results": {
                                            "type": "object",
                                            "example": "Updated Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete a product by its ID",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "brearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "header"
            }
        }
    },
...
}